Fix up hadoop service testing for the nodes.
Annoyingly enough, the nodes might not be fully indexed by the time
we try to find out what roles are on what machines.  We will have to
spin on finding nodes until everything is indexed.

Equally annoying, the hadoop service machinery has an annoying tendency
to report that the pidfile exists but the service is dead even when the
service is in fact alive and well.  Switch to grepping the output of
ps to check for service liveness instead.